requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
imbalanced-learn==0.12.3
mlflow==2.14.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pyyaml==6.0.2
pytest==8.3.2